# perf: calculate anzu count faster and with less memory